+++ iem-mixer/iem-ui/src/pages/mixer/mod.rs
@@ -22,52 +22,11 @@ use crate::components::snapshot_modal::SnapshotModal;
 use crate::components::talk_button::TalkState;
 use crate::components::toolbar::Toolbar;
 
-/// Post-release guard duration in milliseconds.
-/// With server-side echo suppression, this only needs to cover WebSocket round-trip (~10-20ms).
-const POST_RELEASE_GUARD_MS: i32 = 100;
+mod helpers;
+mod push;
 
-/// Minimum interval between WebSocket sends per track (ms).
-/// Limits to ~20 commands/sec to avoid overwhelming the server.
-const THROTTLE_INTERVAL_MS: f64 = 50.0;
-
-/// Processed channel for display (handles stereo pairs)
-/// Note: level_db, pan, muted are read via derived signals from channels
-#[derive(Debug, Clone, PartialEq)]
-struct DisplayChannel {
-    track_index: usize,
-    display_name: String,
-    is_stereo: bool,
-    partner_index: Option<usize>,
-    is_my_input: bool,
-}
-
-/// Send a command via WebSocket (synchronous, non-blocking)
-fn ws_send(ws: ReadSignal<Option<web_sys::WebSocket>>, cmd: &iem_core::ClientMsg) {
-    if let Some(ws) = ws.get_untracked() {
-        if ws.ready_state() == web_sys::WebSocket::OPEN {
-            if let Ok(json) = serde_json::to_string(cmd) {
-                let _ = ws.send_with_str(&json);
-            }
-        }
-    }
-}
-
-/// Storage for WebSocket closures to prevent memory leaks on reconnect.
-/// Dropping a Closure that was passed to JS via `as_ref().unchecked_ref()` properly
-/// releases the WASM-side allocation. Without this, `Closure::forget()` leaks on every reconnect.
-/// Uses Rc<RefCell<>> because wasm_bindgen::Closure is !Send (WASM is single-threaded).
-type WsClosures = (
-    Closure<dyn FnMut(web_sys::MessageEvent)>,
-    Closure<dyn FnMut(web_sys::CloseEvent)>,
-);
-type WsClosureStore = std::rc::Rc<std::cell::RefCell<Option<WsClosures>>>;
-
-/// Counter for consecutive WebSocket failures without receiving data.
-/// Shared across connect_websocket calls via Rc<Cell<>>.
-type WsFailCounter = std::rc::Rc<std::cell::Cell<u32>>;
-
-/// Max consecutive WS failures before redirecting to login
-const MAX_WS_FAILURES: u32 = 3;
+use helpers::*;
+use push::subscribe_to_push;
 
 /// Create and connect a WebSocket, wiring up message handlers to signals
 #[allow(clippy::too_many_arguments)]
@@ -423,225 +382,6 @@ fn connect_websocket(
     // Store closures so they stay alive (preventing JS callback invalidation)
     // and get dropped on next reconnect (preventing memory leak from Closure::forget)
     *ws_closures.borrow_mut() = Some((onmessage, onclose));
-}
-
-/// Subscribe to Web Push for engineer SOS alerts (#133).
-/// Fetches VAPID key, subscribes via Push API, sends subscription to server.
-fn subscribe_to_push() {
-    wasm_bindgen_futures::spawn_local(async move {
-        // 1. Fetch VAPID public key from server
-        let token = match crate::auth::get_token() {
-            Some(t) => t,
-            None => {
-                web_sys::console::log_1(&"[push] no auth token, skipping".into());
-                return;
-            }
-        };
-
-        let resp = match gloo_net::http::Request::get("/api/push/vapid-key")
-            .send()
-            .await
-        {
-            Ok(r) if r.ok() => r,
-            Ok(r) => {
-                web_sys::console::warn_1(
-                    &format!("[push] vapid-key request failed: {}", r.status()).into(),
-                );
-                return;
-            }
-            Err(e) => {
-                web_sys::console::warn_1(&format!("[push] vapid-key fetch error: {:?}", e).into());
-                return;
-            }
-        };
-        let json: serde_json::Value = match resp.json().await {
-            Ok(j) => j,
-            Err(e) => {
-                web_sys::console::warn_1(&format!("[push] vapid-key parse error: {:?}", e).into());
-                return;
-            }
-        };
-        let vapid_key = match json.get("key").and_then(|k| k.as_str()) {
-            Some(k) => k.to_string(),
-            None => {
-                web_sys::console::log_1(&"[push] VAPID not configured, skipping".into());
-                return;
-            }
-        };
-        web_sys::console::log_1(&format!("[push] got VAPID key: {}...", &vapid_key[..20]).into());
-
-        // 2. Get ServiceWorkerRegistration
-        let window = match web_sys::window() {
-            Some(w) => w,
-            None => return,
-        };
-        let navigator = window.navigator();
-        let sw_container: web_sys::ServiceWorkerContainer = match js_sys::Reflect::get(
-            &navigator,
-            &wasm_bindgen::JsValue::from_str("serviceWorker"),
-        )
-        .ok()
-        .and_then(|v| v.dyn_into().ok())
-        {
-            Some(c) => c,
-            None => {
-                web_sys::console::log_1(&"[push] serviceWorker not available, skipping".into());
-                return;
-            }
-        };
-
-        let ready_promise = match sw_container.ready() {
-            Ok(p) => p,
-            Err(e) => {
-                web_sys::console::warn_1(&format!("[push] sw.ready() failed: {:?}", e).into());
-                return;
-            }
-        };
-        web_sys::console::log_1(&"[push] waiting for SW ready...".into());
-        let registration: web_sys::ServiceWorkerRegistration =
-            match wasm_bindgen_futures::JsFuture::from(ready_promise).await {
-                Ok(r) => match r.dyn_into() {
-                    Ok(reg) => reg,
-                    Err(e) => {
-                        web_sys::console::warn_1(
-                            &format!("[push] SW registration cast failed: {:?}", e).into(),
-                        );
-                        return;
-                    }
-                },
-                Err(e) => {
-                    web_sys::console::warn_1(
-                        &format!("[push] SW ready await failed: {:?}", e).into(),
-                    );
-                    return;
-                }
-            };
-        web_sys::console::log_1(&"[push] SW ready, getting push manager...".into());
-
-        // 3. Subscribe to push
-        let push_manager = match registration.push_manager() {
-            Ok(pm) => pm,
-            Err(e) => {
-                web_sys::console::warn_1(&format!("[push] push_manager() failed: {:?}", e).into());
-                return;
-            }
-        };
-
-        // Unsubscribe any existing push subscription first (required when VAPID key changes,
-        // otherwise Chrome rejects subscribe() with a different applicationServerKey)
-        if let Ok(existing_promise) = push_manager.get_subscription() {
-            if let Ok(existing_val) = wasm_bindgen_futures::JsFuture::from(existing_promise).await {
-                if !existing_val.is_null() && !existing_val.is_undefined() {
-                    if let Ok(existing_sub) = existing_val.dyn_into::<web_sys::PushSubscription>() {
-                        let _ = wasm_bindgen_futures::JsFuture::from(
-                            existing_sub.unsubscribe().unwrap_or_else(|_| {
-                                js_sys::Promise::resolve(&wasm_bindgen::JsValue::TRUE)
-                            }),
-                        )
-                        .await;
-                        web_sys::console::log_1(
-                            &"[push] unsubscribed old push subscription".into(),
-                        );
-                    }
-                }
-            }
-        }
-
-        // Decode base64url VAPID key to Uint8Array
-        let key_bytes = match base64url_decode(&vapid_key) {
-            Some(b) => b,
-            None => return,
-        };
-        let key_array = js_sys::Uint8Array::new_with_length(key_bytes.len() as u32);
-        key_array.copy_from(&key_bytes);
-
-        let opts = web_sys::PushSubscriptionOptionsInit::new();
-        opts.set_user_visible_only(true);
-        opts.set_application_server_key(&key_array.into());
-
-        let sub_promise = match push_manager.subscribe_with_options(&opts) {
-            Ok(p) => p,
-            Err(e) => {
-                web_sys::console::warn_1(&format!("[push] subscribe failed: {:?}", e).into());
-                return;
-            }
-        };
-        web_sys::console::log_1(&"[push] subscribing to push...".into());
-        let sub: web_sys::PushSubscription = match wasm_bindgen_futures::JsFuture::from(sub_promise)
-            .await
-        {
-            Ok(v) => match v.dyn_into() {
-                Ok(s) => s,
-                Err(e) => {
-                    web_sys::console::warn_1(
-                        &format!("[push] subscription cast failed: {:?}", e).into(),
-                    );
-                    return;
-                }
-            },
-            Err(e) => {
-                web_sys::console::warn_1(&format!("[push] subscribe await failed: {:?}", e).into());
-                return;
-            }
-        };
-
-        // 4. Send subscription JSON to server
-        let sub_json = match sub.to_json() {
-            Ok(j) => j,
-            Err(_) => return,
-        };
-        let json_str = match js_sys::JSON::stringify(&sub_json)
-            .ok()
-            .and_then(|s| s.as_string())
-        {
-            Some(s) => s,
-            None => return,
-        };
-
-        // Parse the JSON string to a serde_json::Value for gloo_net
-        let body: serde_json::Value = match serde_json::from_str(&json_str) {
-            Ok(v) => v,
-            Err(_) => return,
-        };
-
-        let req = match gloo_net::http::Request::post("/api/push/subscribe")
-            .header("Authorization", &format!("Bearer {}", token))
-            .json(&body)
-        {
-            Ok(r) => r,
-            Err(e) => {
-                web_sys::console::warn_1(&format!("[push] serialize error: {:?}", e).into());
-                return;
-            }
-        };
-        match req.send().await {
-            Ok(r) if r.ok() => {
-                web_sys::console::log_1(&"[push] engineer subscribed to Web Push".into());
-            }
-            Ok(r) => {
-                web_sys::console::warn_1(
-                    &format!("[push] subscribe POST failed: {}", r.status()).into(),
-                );
-            }
-            Err(e) => {
-                web_sys::console::warn_1(&format!("[push] subscribe POST error: {:?}", e).into());
-            }
-        }
-    });
-}
-
-/// Decode base64url (no padding) to bytes.
-/// Note: atob() returns a Latin-1 string (each char = one byte 0-255).
-/// Rust's `.bytes()` gives UTF-8 which mangles values > 127. Use `.chars() as u8` instead.
-fn base64url_decode(input: &str) -> Option<Vec<u8>> {
-    let mut s = input.replace('-', "+").replace('_', "/");
-    while s.len() % 4 != 0 {
-        s.push('=');
-    }
-    web_sys::window()?
-        .atob(&s)
-        .ok()
-        .map(|decoded| decoded.chars().map(|c| c as u8).collect())
 }
 
 /// Mixer page for a specific member
@@ -2816,50 +2556,5 @@ fn ChannelList(
             <Show when=move || open_menu.get().is_some() fallback=|| ()>
                 <div class="ch-menu-backdrop" on:click=move |_| { let _ = set_open_menu.try_set(None); }></div>
             </Show>
-    }
-}
-
-/// Parse track name into main and type parts
-fn parse_track_name(name: &str) -> (String, String) {
-    let parts: Vec<&str> = name.split_whitespace().collect();
-    if parts.len() >= 2 {
-        (parts[0].to_string(), parts[1..].join(" "))
-    } else {
-        (name.to_string(), String::new())
-    }
-}
-
-/// Format dB value for display with unit suffix
-fn format_db(db: f32) -> String {
-    if db <= -60.0 {
-        "-\u{221E}dB".to_string()
-    } else if db >= 0.0 {
-        format!("+{:.1}dB", db)
-    } else {
-        format!("{:.1}dB", db)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_format_db_uses_proper_notation() {
-        assert!(format_db(0.0).ends_with("dB"), "Must use 'dB' not 'db'");
-        assert!(format_db(-6.0).ends_with("dB"));
-        assert!(format_db(-60.0).ends_with("dB")); // -inf case
-    }
-
-    #[test]
-    fn test_format_db_max_length() {
-        let cases = [0.0, 6.0, 12.0, -6.0, -12.5, -59.9, -60.0, -100.0];
-        for db in cases {
-            let s = format_db(db);
-            assert!(
-                s.chars().count() <= 7,
-                "format_db({db}) = \"{s}\" exceeds 7 chars"
-            );
-        }
     }
 }
